+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Release Notes
 
 ## 0.6.1
-- Added companion Dialog - outline
+- NEW! Companion dialog which shows which actors have saved tokens and which will be generated
 
 ## 0.6.0abc
 - Allow Compendium entries as well as Actors
-- Fixed? Combat tracker now pops up more reliably
-- Was popping an extra copy of the tutorial every time you pressed the button
+- Combat tracker now pops up more reliably when run a Quick Encounter
+- No longer pops up an extra copy of the tutorial every time you press the Quick Encounter button
 - Handle multipliers that are dice rolls, e.g. 1d4+2 Vampire Spawn (Note they must be in Foundry [[/r 1d4+2]] form  to be recognized)
 
 ## v0.5.5